[#4745] Clean up the duplicate code associated with the Zookeeper
Clean up the duplicate code associated with the Zookeeper
