+++ collector/src/main/java/com/navercorp/pinpoint/collector/cluster/flink/FlinkClusterService.java
@@ -15,8 +15,12 @@
  */
 package com.navercorp.pinpoint.collector.cluster.flink;
 
-import com.navercorp.pinpoint.collector.cluster.zookeeper.*;
+import com.navercorp.pinpoint.collector.cluster.zookeeper.DefaultZookeeperClient;
+import com.navercorp.pinpoint.collector.cluster.zookeeper.ZookeeperClient;
+import com.navercorp.pinpoint.collector.cluster.zookeeper.ZookeeperClusterManager;
+import com.navercorp.pinpoint.collector.cluster.zookeeper.ZookeeperUtils;
 import com.navercorp.pinpoint.collector.config.CollectorConfiguration;
+import com.navercorp.pinpoint.common.server.cluster.zookeeper.ZookeeperEventWatcher;
 import com.navercorp.pinpoint.common.server.util.concurrent.CommonState;
 import com.navercorp.pinpoint.common.server.util.concurrent.CommonStateContext;
 import com.navercorp.pinpoint.common.util.Assert;
