+++ collector/src/main/java/com/navercorp/pinpoint/collector/cluster/zookeeper/ZookeeperClient.java
@@ -16,7 +16,7 @@
 
 package com.navercorp.pinpoint.collector.cluster.zookeeper;
 
-import com.navercorp.pinpoint.collector.cluster.zookeeper.exception.PinpointZookeeperException;
+import com.navercorp.pinpoint.common.server.cluster.zookeeper.exception.PinpointZookeeperException;
 
 import java.io.IOException;
 import java.util.List;
