+++ commons-server/src/main/java/com/navercorp/pinpoint/common/server/cluster/zookeeper/exception/AuthException.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2017 NAVER Corp.
+ * Copyright 2018 NAVER Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,13 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.navercorp.pinpoint.web.cluster.zookeeper;
 
-import org.apache.zookeeper.Watcher;
+package com.navercorp.pinpoint.common.server.cluster.zookeeper.exception;
 
 /**
- * @author minwoo.jung
+ * @author koo.taejin
  */
-public interface ZookeeperEventWatcher extends Watcher {
-    boolean isConnected();
+public class AuthException extends PinpointZookeeperException {
+
+    public AuthException() {
+    }
+
+    public AuthException(String message) {
+        super(message);
+    }
+
+    public AuthException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public AuthException(Throwable cause) {
+        super(cause);
+    }
+
 }
