+++ commons-server/src/main/java/com/navercorp/pinpoint/common/server/cluster/zookeeper/exception/ConnectionException.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2014 NAVER Corp.
+ * Copyright 2018 NAVER Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,25 +14,25 @@
  * limitations under the License.
  */
 
-package com.navercorp.pinpoint.collector.cluster.zookeeper.exception;
+package com.navercorp.pinpoint.common.server.cluster.zookeeper.exception;
 
 /**
  * @author koo.taejin
  */
-public class AuthException extends PinpointZookeeperException {
+public class ConnectionException extends PinpointZookeeperException {
 
-    public AuthException() {
+    public ConnectionException() {
     }
 
-    public AuthException(String message) {
+    public ConnectionException(String message) {
         super(message);
     }
 
-    public AuthException(String message, Throwable cause) {
+    public ConnectionException(String message, Throwable cause) {
         super(message, cause);
     }
 
-    public AuthException(Throwable cause) {
+    public ConnectionException(Throwable cause) {
         super(cause);
     }
