+++ commons-server/src/main/java/com/navercorp/pinpoint/common/server/cluster/zookeeper/exception/UnknownException.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2014 NAVER Corp.
+ * Copyright 2018 NAVER Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,15 +14,26 @@
  * limitations under the License.
  */
 
-package com.navercorp.pinpoint.collector.cluster.zookeeper;
-
-import org.apache.zookeeper.Watcher;
+package com.navercorp.pinpoint.common.server.cluster.zookeeper.exception;
 
 /**
  * @author koo.taejin
  */
-public interface ZookeeperEventWatcher extends Watcher {
+public class UnknownException extends PinpointZookeeperException {
 
-    boolean isConnected();
+    public UnknownException() {
+    }
+
+    public UnknownException(String message) {
+        super(message);
+    }
+
+    public UnknownException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public UnknownException(Throwable cause) {
+        super(cause);
+    }
 
 }
