+++ web/src/main/java/com/navercorp/pinpoint/web/cluster/zookeeper/ZookeeperClient.java
@@ -17,15 +17,15 @@
 package com.navercorp.pinpoint.web.cluster.zookeeper;
 
 
+import com.navercorp.pinpoint.common.server.cluster.zookeeper.ZookeeperConnectionManager;
+import com.navercorp.pinpoint.common.server.cluster.zookeeper.ZookeeperConstatns;
+import com.navercorp.pinpoint.common.server.cluster.zookeeper.ZookeeperEventWatcher;
+import com.navercorp.pinpoint.common.server.cluster.zookeeper.ZookeeperExceptionResolver;
+import com.navercorp.pinpoint.common.server.cluster.zookeeper.exception.ConnectionException;
+import com.navercorp.pinpoint.common.server.cluster.zookeeper.exception.PinpointZookeeperException;
+import com.navercorp.pinpoint.common.server.cluster.zookeeper.exception.UnknownException;
 import com.navercorp.pinpoint.common.server.util.concurrent.CommonStateContext;
 import com.navercorp.pinpoint.rpc.util.TimerFactory;
-import com.navercorp.pinpoint.web.cluster.zookeeper.exception.AuthException;
-import com.navercorp.pinpoint.web.cluster.zookeeper.exception.BadOperationException;
-import com.navercorp.pinpoint.web.cluster.zookeeper.exception.ConnectionException;
-import com.navercorp.pinpoint.web.cluster.zookeeper.exception.NoNodeException;
-import com.navercorp.pinpoint.web.cluster.zookeeper.exception.PinpointZookeeperException;
-import com.navercorp.pinpoint.web.cluster.zookeeper.exception.TimeoutException;
-import com.navercorp.pinpoint.web.cluster.zookeeper.exception.UnknownException;
 import org.apache.zookeeper.CreateMode;
 import org.apache.zookeeper.KeeperException;
 import org.apache.zookeeper.KeeperException.Code;
@@ -44,7 +44,6 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 /**
- *
  * <strong>Conditional thread safe</strong> <br>
  * If multiple threads invokes connect, reconnect, or close concurrently; then it is possible for the object's zookeeper field to be out of sync.
  *
@@ -53,40 +52,30 @@ import java.util.concurrent.TimeUnit;
 public class ZookeeperClient {
 
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
-    public static final long DEFAULT_RECONNECT_DELAY_WHEN_SESSION_EXPIRED = 30000;
 
-    private final CommonStateContext stateContext;
+    private final CommonStateContext stateContext = new CommonStateContext();
 
     private final HashedWheelTimer timer;
 
-    private final String hostPort;
-    private final int sessionTimeout;
-    private final ZookeeperEventWatcher zookeeperEventWatcher;
-    private final long reconnectDelayWhenSessionExpired;
+    private final ZookeeperConnectionManager zookeeperConnectionManager;
 
-    // ZK client is thread-safe
-    private volatile ZooKeeper zookeeper;
+    private final long reconnectDelayWhenSessionExpired;
 
     // hmm this structure should contain all necessary information
     public ZookeeperClient(String hostPort, int sessionTimeout, ZookeeperEventWatcher manager) {
-        this(hostPort, sessionTimeout, manager, DEFAULT_RECONNECT_DELAY_WHEN_SESSION_EXPIRED);
+        this(hostPort, sessionTimeout, manager, ZookeeperConstatns.DEFAULT_RECONNECT_DELAY_WHEN_SESSION_EXPIRED);
     }
-    
+
     public ZookeeperClient(String hostPort, int sessionTimeout, ZookeeperEventWatcher zookeeperEventWatcher, long reconnectDelayWhenSessionExpired) {
-        this.hostPort = hostPort;
-        this.sessionTimeout = sessionTimeout;
-        this.zookeeperEventWatcher = zookeeperEventWatcher;
+        this.zookeeperConnectionManager = new ZookeeperConnectionManager(hostPort, sessionTimeout, zookeeperEventWatcher);
+
         this.reconnectDelayWhenSessionExpired = reconnectDelayWhenSessionExpired;
-        
-
         this.timer = TimerFactory.createHashedWheelTimer(this.getClass().getSimpleName(), 100, TimeUnit.MILLISECONDS, 512);
-
-        this.stateContext = new CommonStateContext();
     }
 
     public void connect() throws IOException {
         if (stateContext.changeStateInitializing()) {
-            this.zookeeper = new ZooKeeper(hostPort, sessionTimeout, zookeeperEventWatcher); // server
+            zookeeperConnectionManager.start();
             stateContext.changeStateStarted();
         } else {
             logger.warn("connect() failed. error : Illegal State. State may be {}.", stateContext.getCurrentState());
@@ -99,49 +88,21 @@ public class ZookeeperClient {
             return;
         }
 
-        ZooKeeper zookeeper = this.zookeeper;
-        if (zookeeper.getState().isConnected()) {
-            logger.warn("ZookeeperClient.reconnectWhenSessionExpired() failed. Error: session(0x{}) is connected.", Long.toHexString(zookeeper.getSessionId()));
-            return;
-        }
-
-        logger.warn("Execute ZookeeperClient.reconnectWhenSessionExpired()(Expired session:0x{}).", Long.toHexString(zookeeper.getSessionId()));
-
-        try {
-            zookeeper.close();
-        } catch (InterruptedException e) {
-            Thread.currentThread().interrupt();
-        }
-        
-        ZooKeeper newZookeeper = createNewZookeeper();
-        if (newZookeeper == null) {
-            logger.warn("Failed to create new Zookeeper instance. It will be retry  AFTER {}ms.", reconnectDelayWhenSessionExpired);
-
+        boolean hasConnectedZookeeper = zookeeperConnectionManager.reconnectWhenSessionExpired();
+        if (!hasConnectedZookeeper) {
             timer.newTimeout(new TimerTask() {
                 @Override
                 public void run(Timeout timeout) throws Exception {
                     if (timeout.isCancelled()) {
                         return;
                     }
-                    
+
                     reconnectWhenSessionExpired();
                 }
             }, reconnectDelayWhenSessionExpired, TimeUnit.MILLISECONDS);
-        } else {
-            this.zookeeper = newZookeeper;
         }
-
     }
 
-    private ZooKeeper createNewZookeeper() {
-        try {
-            return new ZooKeeper(hostPort, sessionTimeout, zookeeperEventWatcher);
-        } catch (IOException ignore) {
-            // ignore
-        }
-        return null;
-    }
-    
     /**
      * do not create node in path suffix
      *
@@ -151,7 +112,7 @@ public class ZookeeperClient {
     public void createPath(String path) throws PinpointZookeeperException, InterruptedException {
         checkState();
 
-        ZooKeeper zookeeper = this.zookeeper;
+        ZooKeeper zookeeper = zookeeperConnectionManager.getZookeeper();
         int pos = 1;
         do {
             pos = path.indexOf('/', pos + 1);
@@ -166,7 +127,7 @@ public class ZookeeperClient {
                     zookeeper.create(subPath, new byte[0], Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
                 } catch (KeeperException exception) {
                     if (exception.code() != Code.NODEEXISTS) {
-                        handleException(exception);
+                        ZookeeperExceptionResolver.resolve(exception, true);
                     }
                 }
             } else {
@@ -178,8 +139,8 @@ public class ZookeeperClient {
     // we need deep node inspection for verification purpose (node content)
     public String createNode(String zNodePath, byte[] data, CreateMode createMode) throws PinpointZookeeperException, InterruptedException {
         checkState();
-        
-        ZooKeeper zookeeper = this.zookeeper;
+
+        ZooKeeper zookeeper = zookeeperConnectionManager.getZookeeper();
         try {
             if (zookeeper.exists(zNodePath, false) != null) {
                 return zNodePath;
@@ -189,7 +150,7 @@ public class ZookeeperClient {
             return pathName;
         } catch (KeeperException exception) {
             if (exception.code() != Code.NODEEXISTS) {
-                handleException(exception);
+                ZookeeperExceptionResolver.resolve(exception, true);
             }
         }
         return zNodePath;
@@ -198,12 +159,12 @@ public class ZookeeperClient {
     public List<String> getChildren(String path, boolean watch) throws PinpointZookeeperException, InterruptedException {
         checkState();
 
-        ZooKeeper zookeeper = this.zookeeper;
+        ZooKeeper zookeeper = zookeeperConnectionManager.getZookeeper();
         try {
             return zookeeper.getChildren(path, watch);
         } catch (KeeperException exception) {
             if (exception.code() != Code.NONODE) {
-                handleException(exception);
+                ZookeeperExceptionResolver.resolve(exception, true);
             }
         }
 
@@ -217,11 +178,11 @@ public class ZookeeperClient {
     public byte[] getData(String path, boolean watch) throws PinpointZookeeperException, InterruptedException {
         checkState();
 
-        ZooKeeper zookeeper = this.zookeeper;
+        ZooKeeper zookeeper = zookeeperConnectionManager.getZookeeper();
         try {
             return zookeeper.getData(path, watch, null);
         } catch (KeeperException exception) {
-            handleException(exception);
+            ZookeeperExceptionResolver.resolve(exception, true);
         }
 
         throw new UnknownException("UnknownException.");
@@ -231,12 +192,12 @@ public class ZookeeperClient {
     public void delete(String path) throws PinpointZookeeperException, InterruptedException {
         checkState();
 
-        ZooKeeper zookeeper = this.zookeeper;
+        ZooKeeper zookeeper = zookeeperConnectionManager.getZookeeper();
         try {
             zookeeper.delete(path, -1);
         } catch (KeeperException exception) {
             if (exception.code() != Code.NONODE) {
-                handleException(exception);
+                ZookeeperExceptionResolver.resolve(exception, true);
             }
         }
     }
@@ -244,7 +205,7 @@ public class ZookeeperClient {
     public boolean exists(String path) throws PinpointZookeeperException, InterruptedException {
         checkState();
 
-        ZooKeeper zookeeper = this.zookeeper;
+        ZooKeeper zookeeper = zookeeperConnectionManager.getZookeeper();
         try {
             Stat stat = zookeeper.exists(path, false);
             if (stat == null) {
@@ -252,40 +213,16 @@ public class ZookeeperClient {
             }
         } catch (KeeperException exception) {
             if (exception.code() != Code.NODEEXISTS) {
-                handleException(exception);
+                ZookeeperExceptionResolver.resolve(exception, true);
             }
         }
         return true;
     }
 
     private void checkState() throws PinpointZookeeperException {
-        if (!zookeeperEventWatcher.isConnected() || !stateContext.isStarted()) {
+        if (!zookeeperConnectionManager.isConnected() || !stateContext.isStarted()) {
             throw new ConnectionException("Instance must be connected.");
         }
-    }
-
-    private void handleException(KeeperException keeperException) throws PinpointZookeeperException {
-        switch (keeperException.code()) {
-            case CONNECTIONLOSS:
-            case SESSIONEXPIRED:
-                throw new ConnectionException(keeperException.getMessage(), keeperException);
-            case AUTHFAILED:
-            case INVALIDACL:
-            case NOAUTH:
-                throw new AuthException(keeperException.getMessage(), keeperException);
-            case BADARGUMENTS:
-            case BADVERSION:
-            case NOCHILDRENFOREPHEMERALS:
-            case NOTEMPTY:
-            case NODEEXISTS:
-                throw new BadOperationException(keeperException.getMessage(), keeperException);
-            case NONODE:
-                throw new NoNodeException(keeperException.getMessage(), keeperException);
-            case OPERATIONTIMEOUT:
-                throw new TimeoutException(keeperException.getMessage(), keeperException);
-            default:
-                throw new UnknownException(keeperException.getMessage(), keeperException);
-            }
     }
 
     public void close() {
@@ -294,13 +231,7 @@ public class ZookeeperClient {
                 timer.stop();
             }
 
-            if (zookeeper != null) {
-                try {
-                    zookeeper.close();
-                } catch (InterruptedException ignore) {
-                    logger.debug(ignore.getMessage(), ignore);
-                }
-            }
+            zookeeperConnectionManager.stop();
             stateContext.changeStateStopped();
         } else {
             logger.warn("close failed. error : Illegal State. State may be {}.", stateContext.getCurrentState());
